+++ x/common/types/packet_uid.go
@@ -1,17 +1,17 @@
 package types
 
-import fmt "fmt"
+import "fmt"
 
 // PacketUID is a unique identifier for an Rollapp IBC packet on the hub
 type PacketUID struct {
-	Type              RollappPacket_Type
+	Type              Type
 	RollappHubPort    string
 	RollappHubChannel string
 	Sequence          uint64
 }
 
 // NewPacketUID creates a new PacketUID with the provided details.
-func NewPacketUID(packetType RollappPacket_Type, hubPort string, hubChannel string, sequence uint64) PacketUID {
+func NewPacketUID(packetType Type, hubPort string, hubChannel string, sequence uint64) PacketUID {
 	return PacketUID{
 		Type:              packetType,
 		RollappHubPort:    hubPort,