+++ x/delayedack/keeper/grpc_query.go
@@ -4,6 +4,8 @@ import (
 	"context"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	commontypes "github.com/dymensionxyz/dymension/v3/x/common/types"
 	"github.com/dymensionxyz/dymension/v3/x/delayedack/types"
 
 	"google.golang.org/grpc/codes"
@@ -40,10 +42,14 @@ func (q Querier) GetPackets(goCtx context.Context, req *types.QueryRollappPacket
 	res := &types.QueryRollappPacketListResponse{}
 
 	if req.RollappId == "" {
-		// query by status
-		res.RollappPackets = q.ListRollappPackets(ctx, types.ByStatus(req.Status))
+		// query by status (PENDING by default)
+		if req.Type != commontypes.Type_UNDEFINED {
+			res.RollappPackets = q.ListRollappPackets(ctx, types.ByType(req.Type))
+		} else {
+			res.RollappPackets = q.ListRollappPackets(ctx, types.ByStatus(req.Status))
+		}
 	} else {
-		// query by rollapp id
+		// query by rollapp id and status (PENDING by default)
 		res.RollappPackets = q.ListRollappPackets(ctx, types.ByRollappIDByStatus(req.RollappId, req.Status))
 	}