+++ x/delayedack/proof_height_ante.go
@@ -5,6 +5,7 @@ import (
 
 	clienttypes "github.com/cosmos/ibc-go/v6/modules/core/02-client/types"
 	channeltypes "github.com/cosmos/ibc-go/v6/modules/core/04-channel/types"
+
 	commontypes "github.com/dymensionxyz/dymension/v3/x/common/types"
 	delayedacktypes "github.com/dymensionxyz/dymension/v3/x/delayedack/types"
 )
@@ -25,7 +26,7 @@ func (rrd IBCProofHeightDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simula
 		case *channeltypes.MsgRecvPacket:
 			height = msg.ProofHeight
 			packetId = commontypes.NewPacketUID(
-				commontypes.RollappPacket_ON_RECV,
+				commontypes.Type_ON_RECV,
 				msg.Packet.DestinationPort,
 				msg.Packet.DestinationChannel,
 				msg.Packet.Sequence,
@@ -34,7 +35,7 @@ func (rrd IBCProofHeightDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simula
 		case *channeltypes.MsgAcknowledgement:
 			height = msg.ProofHeight
 			packetId = commontypes.NewPacketUID(
-				commontypes.RollappPacket_ON_ACK,
+				commontypes.Type_ON_ACK,
 				msg.Packet.SourcePort,
 				msg.Packet.SourceChannel,
 				msg.Packet.Sequence,
@@ -43,7 +44,7 @@ func (rrd IBCProofHeightDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simula
 		case *channeltypes.MsgTimeout:
 			height = msg.ProofHeight
 			packetId = commontypes.NewPacketUID(
-				commontypes.RollappPacket_ON_TIMEOUT,
+				commontypes.Type_ON_TIMEOUT,
 				msg.Packet.SourcePort,
 				msg.Packet.SourceChannel,
 				msg.Packet.Sequence,